finishing part 4 (LFG)!! only fixed integ tests for blogs

+++ part4/blog-list/controllers/blogs.js
@@ -1,31 +1,18 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
-const jwt = require('jsonwebtoken')
+// const User = require('../models/user')
+// const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
 
 
 blogRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
 })
 
-const getTokenFrom = request => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-        return authorization.substring(7)
-    }
-    return null
-}
-
-blogRouter.post('/', async (request, response) => {
+blogRouter.post('/', middleware.userExtractor, async (request, response) => {
     const body = request.body
-    const token = getTokenFrom(body)
-    const decoodedToken = jwt.verify(token, process.env.SECRET)
-    if (!decoodedToken.id) {
-        return response.status(401).json({ error: 'token missing or invalid' })
-    }
-
-    const user = await User.findById(decoodedToken.id)
+    const user = request.user
 
     const blog = new Blog({
         title: body.title,
@@ -36,10 +23,20 @@ blogRouter.post('/', async (request, response) => {
     })
 
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    user.save()
     response.status(201).json(savedBlog)
 })
 
-blogRouter.delete('/:id', async (request, response) => {
+blogRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+    const user = request.user
+    if (!user) {
+        return response.status(404).json({ error: 'user no longer exists' })
+    }
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() !== user.id.toString()) {
+        return response.status(403).json({ error: 'user can only delete their own blog posts' })
+    }
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
